added Makefile.generic files for tests
git-svn-id: http://dyncall.org/svn/dyncall/trunk/dyncall@895 c71cb566-5a31-4692-9165-cf76bdc42cab

+++ test/nm/nm.c
@@ -31,7 +31,7 @@ void list_syms(DLLib* pLib)
   for (; i < n; ++i) {
     const char* name = dlSymsName(pSyms,i);
     void* addr = dlSymsValue(pSyms,i);
-    printf("%s %lx\n", name, addr);
+    printf("%s %lx\n", name, (ptrdiff_t) addr);
   }
   dlSymsCleanup(pSyms);
   free(pSyms);
